Simplify the module architecture
